rename directory to backend
